refactor: modularize main
